# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SoulForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # PLAYER
...
  /bosses:
    get:
      operationId: listBosses
      tags: [bosses]
      summary: List available bosses
      responses:
        "200":
          description: Boss list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Boss"
    post:
      operationId: createCustomBoss
      tags: [bosses]
      summary: Create a custom boss (pro feature)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BossInput"
      responses:
        "201":
          description: Created boss
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Boss"
  /bosses/active:
    get:
      operationId: getActiveBoss
      tags: [bosses]
      summary: Get currently challenged boss
      responses:
        "200":
          description: Active boss battle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BossBattle"
        "404":
          description: No active boss
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bosses/{bossId}/challenge:
    post:
      operationId: challengeBoss
      tags: [bosses]
      summary: Start challenging a boss
      parameters:
        - name: bossId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Boss battle started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BossBattle"

  # AI CHALLENGES
...
  /calendar:
    get:
      operationId: getCalendar
      tags: [calendar]
      summary: Get calendar entries for a month
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: month
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Calendar data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarDay"
  /calendar/{date}:
    get:
      operationId: getCalendarDay
      tags: [calendar]
      summary: Get entries for a specific date
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Day entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarDay"

  # DASHBOARD
...
  /admin/bosses:
    get:
      operationId: adminListBosses
      tags: [admin]
      summary: List all bosses
      responses:
        "200":
          description: Boss list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Boss"
    post:
      operationId: adminCreateBoss
      tags: [admin]
      summary: Create a boss
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBossInput"
      responses:
        "201":
          description: Created boss
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Boss"
  /admin/bosses/{bossId}:
    patch:
      operationId: adminUpdateBoss
      tags: [admin]
      summary: Update a boss
      parameters:
        - name: bossId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBossInput"
      responses:
        "200":
          description: Updated boss
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Boss"
    delete:
      operationId: adminDeleteBoss
      tags: [admin]
      summary: Delete a boss
      parameters:
        - name: bossId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # AUTH
    AuthInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
    User:
      type: object
      required: [id, username, role, createdAt, isPro]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string
          enum: [user, admin, moderator]
        createdAt:
          type: string
        isPro:
          type: boolean
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # PLAYER
...
    Boss:
      type: object
      required: [id, name, description, difficulty, maxHp, xpReward, goldReward, lootTable, imageUrl, isCustom, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, legendary]
        maxHp:
          type: integer
        xpReward:
          type: integer
        goldReward:
          type: integer
        lootTable:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
        isCustom:
          type: boolean
        createdAt:
          type: string
    BossInput:
      type: object
      required: [name, description, difficulty, maxHp]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, legendary]
        maxHp:
          type: integer
        xpReward:
          type: integer
        goldReward:
          type: integer
    BossBattle:
      type: object
      required: [id, userId, boss, currentHp, startedAt, attackPoints]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        boss:
          $ref: "#/components/schemas/Boss"
        currentHp:
          type: integer
        startedAt:
          type: string
        attackPoints:
          type: integer
...
    CalendarDay:
      type: object
      required: [date, habitsCompleted, habitsTotal, subtasksCompleted, xpEarned, hasCheckin]
      properties:
        date:
          type: string
        habitsCompleted:
          type: integer
        habitsTotal:
          type: integer
        subtasksCompleted:
          type: integer
        xpEarned:
          type: integer
        hasCheckin:
          type: boolean

    # DASHBOARD
...
    AdminBossInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, legendary]
        maxHp:
          type: integer
        xpReward:
          type: integer
        goldReward:
          type: integer
        lootTable:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
        isCustom:
          type: boolean
...